add ClienteMongo and fix short to int for tentativi accesso

+++ applicazione-bancaria/src/main/java/eu/tasgroup/applicativo/businesscomponent/model/mysql/Amministratore.java
@@ -28,7 +28,7 @@ public class Amministratore implements Serializable {
     private String passwordAdmin;
 
     @Column(name = "tentativi_errati")
-    private short tentativiErrati = 0;
+    private int tentativiErrati = 0;
 
     @Column(name = "account_bloccato", columnDefinition = "tinyint(1) default 0")
     private boolean accountBloccato = false;
@@ -73,11 +73,11 @@ public class Amministratore implements Serializable {
 		this.passwordAdmin = passwordAdmin;
 	}
 
-	public short getTentativiErrati() {
+	public int getTentativiErrati() {
 		return tentativiErrati;
 	}
 
-	public void setTentativiErrati(short tentativiErrati) {
+	public void setTentativiErrati(int tentativiErrati) {
 		this.tentativiErrati = tentativiErrati;
 	}
